+++ changelog.md
@@ -1,7 +1,7 @@
 # The Modding Tree changelog:
 
 # v2.5.3 - 5/8/21
-- Improved performance of bars.
+- Improved performance of tab formats and bars.
 - Respec confirmation settings are now kept on resets.
 - Improved compatibility with older browsers.
 - Fixed missing pixel on vertical bars.